requirements.txt
pandas 
faiss-cpu 
numpy 
openai
streamlit